requirements.txt
altgraph==0.17.4
comtypes==1.2.1
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
psutil==5.9.8
pycaw==20230407
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
pyserial==3.5
pystray==0.19.5
pywin32-ctypes==0.2.2
setuptools==69.0.3
six==1.16.0
